New F1ap but with stack overflow

+++ qcore/src/procedures/ue_procedures/mod.rs
@@ -124,7 +124,7 @@ impl<'a, A: HandlerApi> UeProcedure<'a, A> {
 
     async fn rrc_indication<T: Send + SerDes>(&mut self, srb_id: SrbId, rrc: T) -> Result<()> {
         let rrc_bytes = rrc.into_bytes()?;
-        let rrc_container = maybe_pdcp_encapsulate(rrc_bytes, srb_id.0, &mut self.ue.pdcp_tx);
+        let rrc_container = maybe_pdcp_encapsulate(rrc_bytes, srb_id.0 as u8, &mut self.ue.pdcp_tx);
         let dl_message = crate::f1ap::build::dl_rrc_message_transfer(
             self.ue.key,
             self.ue.gnb_du_ue_f1ap_id,
